chore: RegisterFont 파일 분리
